requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
PyYAML>=6.0.1
requests>=2.31.0
joblib>=1.4.0
fastapi>=0.115.0
uvicorn>=0.30.0
